My first API
Back at it. Due to covid-19 and working on material for #codeLouisville I stepped away from this bootcamp for a moment but picked it back up. I'll be building a weather application with an actual back-end on it. This is step #1, create a server via node.js, call an API for real-time data, and format it in JavaScript Object Notation.
Simple for now, but we'll see what it looks like once we're finished.